[gstreamer] some fixes avoiding gst to work properly

+++ branches/nhd2-exp/src/gui/pluginlist.cpp
@@ -138,7 +138,7 @@ int CPluginList::exec(CMenuTarget * parent, const std::string &)
 	frameBuffer->blit();
 #endif	
 
-	uint64_t timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
+	unsigned long long int timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings::TIMING_MENU]);
 
 	bool loop = true;
 	while (loop)
